Move files under processor/module/ to processor/
