+++ src/main/java/ai/knowly/langtorch/capability/modality/text/TextCompletionTextLLMCapability.java
@@ -4,7 +4,7 @@ import static com.google.common.util.concurrent.Futures.immediateFuture;
 import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 
 import ai.knowly.langtorch.preprocessing.parser.Parser;
-import ai.knowly.langtorch.processor.module.Processor;
+import ai.knowly.langtorch.processor.Processor;
 import ai.knowly.langtorch.schema.text.SingleText;
 import com.google.common.util.concurrent.FluentFuture;
 import com.google.common.util.concurrent.ListenableFuture;
