+++ src/main/java/ai/knowly/langtorch/capability/module/openai/SimpleChatCapability.java
@@ -4,7 +4,7 @@ import ai.knowly.langtorch.capability.modality.text.ChatCompletionLLMCapability;
 import ai.knowly.langtorch.preprocessing.parser.ChatMessageToStringParser;
 import ai.knowly.langtorch.preprocessing.parser.Parser;
 import ai.knowly.langtorch.preprocessing.parser.StringToMultiChatMessageParser;
-import ai.knowly.langtorch.processor.module.openai.chat.OpenAIChatProcessor;
+import ai.knowly.langtorch.processor.openai.chat.OpenAIChatProcessor;
 import ai.knowly.langtorch.schema.chat.ChatMessage;
 import ai.knowly.langtorch.schema.text.MultiChatMessage;
 import ai.knowly.langtorch.store.memory.conversation.ConversationMemory;
