+++ src/main/java/ai/knowly/langtorch/processor/ProcessorConfig.java
@@ -1,0 +1,3 @@
+package ai.knowly.langtorch.processor;
+
+public interface ProcessorConfig {}
