--- src/main/java/ai/knowly/langtorch/processor/module/EmbeddingsProcessor.java
@@ -1,7 +1,0 @@
-package ai.knowly.langtorch.processor.module;
-
-import ai.knowly.langtorch.schema.embeddings.EmbeddingInput;
-import ai.knowly.langtorch.schema.embeddings.EmbeddingOutput;
-
-/** EmbeddingsProcessor is a shared interface for embedding output. */
-public interface EmbeddingsProcessor extends Processor<EmbeddingInput, EmbeddingOutput> {}
