--- src/main/java/ai/knowly/langtorch/processor/module/ProcessorConfig.java
@@ -1,3 +1,0 @@
-package ai.knowly.langtorch.processor.module;
-
-public interface ProcessorConfig {}
