--- src/main/java/ai/knowly/langtorch/processor/module/openai/chat/UnknownMessageException.java
@@ -1,7 +1,0 @@
-package ai.knowly.langtorch.processor.module.openai.chat;
-
-public class UnknownMessageException extends RuntimeException {
-  public UnknownMessageException(String message) {
-    super(message);
-  }
-}
