+++ src/main/java/ai/knowly/langtorch/processor/openai/chat/UnknownMessageException.java
@@ -1,0 +1,7 @@
+package ai.knowly.langtorch.processor.openai.chat;
+
+public class UnknownMessageException extends RuntimeException {
+  public UnknownMessageException(String message) {
+    super(message);
+  }
+}
