+++ src/test/java/ai/knowly/langtorch/capability/module/openai/PromptTemplateCapabilityTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.when;
 
 import ai.knowly.langtorch.llm.openai.OpenAIService;
-import ai.knowly.langtorch.processor.module.openai.text.OpenAITextProcessor;
+import ai.knowly.langtorch.processor.openai.text.OpenAITextProcessor;
 import ai.knowly.langtorch.prompt.template.PromptTemplate;
 import ai.knowly.langtorch.util.OpenAIServiceTestingUtils;
 import com.google.common.collect.ImmutableMap;
