+++ src/test/java/ai/knowly/langtorch/capability/module/openai/SimpleChatCapabilityTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.when;
 
 import ai.knowly.langtorch.llm.openai.OpenAIService;
-import ai.knowly.langtorch.processor.module.openai.chat.OpenAIChatProcessor;
+import ai.knowly.langtorch.processor.openai.chat.OpenAIChatProcessor;
 import ai.knowly.langtorch.schema.chat.AssistantMessage;
 import ai.knowly.langtorch.util.OpenAIServiceTestingUtils;
 import org.junit.jupiter.api.Test;
